Consolidate AgpHandlers
