--- agp-handlers/agp-handler-83/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=sgp-agp-handler-83
-POM_NAME=Slack Gradle Plugin (AGP Handler 8.3)
-POM_DESCRIPTION=Slack Gradle Plugin (AGP Handler 8.3)
