+++ settings.gradle.kts
@@ -179,7 +179,6 @@ rootProject.name = "slack-gradle-plugin"
 
 // Please keep these in alphabetical order!
 include(
-  ":agp-handlers:agp-handler-83",
   ":agp-handlers:agp-handler-api",
   ":skippy",
   ":sgp-common",
