+++ slack-plugin/build.gradle.kts
@@ -72,7 +72,6 @@ dependencies {
   api(libs.okhttp)
   // Better I/O
   api(libs.okio)
-  api(projects.agpHandlers.agpHandler83)
   api(projects.agpHandlers.agpHandlerApi)
 
   implementation(platform(libs.coroutines.bom))
