+++ slack-plugin/lint-baseline.xml
@@ -1,5 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<issues format="6" by="lint 8.4.0-alpha11" type="baseline" client="gradle" dependencies="false" name="AGP (8.3.1)" variant="all" version="8.4.0-alpha11">
+<issues format="6" by="lint 8.4.0-alpha13" type="baseline" client="gradle" dependencies="false" name="AGP (8.3.1)" variant="all" version="8.4.0-alpha13">
+
+    <issue
+        id="InternalGradleApiUsage"
+        message="Avoid using internal Android Gradle Plugin APIs"
+        errorLine1="import com.android.build.gradle.internal.SdkLocationSourceSet"
+        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
+        <location
+            file="src/main/kotlin/slack/gradle/AndroidSourcesConfigurer.kt"
+            line="18"
+            column="1"/>
+    </issue>
+
+    <issue
+        id="InternalGradleApiUsage"
+        message="Avoid using internal Android Gradle Plugin APIs"
+        errorLine1="import com.android.build.gradle.internal.SdkLocator"
+        errorLine2="~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~">
+        <location
+            file="src/main/kotlin/slack/gradle/AndroidSourcesConfigurer.kt"
+            line="19"
+            column="1"/>
+    </issue>
 
     <issue
         id="InternalGradleApiUsage"
